Changed input/output parameter requirements.
ChatterBot now requires a list of input and output adapter pairs
to be provided. Any number of pairs can be added, and each pair
must be a tuple containing only two elements. The first element
must be the dot notated import path to an input adapter and the
second element must be the dot notated import path to an output
adapter.